Update: define logo into env instead and fix comment
