+++ src/constants/common.ts
@@ -1,2 +1,1 @@
 export const formatDate = 'DD/MM/YYYY'
-export const company = 'Company'
