+++ src/constants/env.ts
@@ -1,0 +1,4 @@
+const ENV = {
+  LOGO_COMPANY: import.meta.env.VITE_API_COMPANY_NAME,
+}
+export default ENV
